pyproject.toml
[tool.poetry]
name = "repl_python3_NotBot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
pil-compat = "^1.0.0"
Flask = "^2.0.2"
discord-buttons-plugin = "^0.1.2"
discord-ui = "^5.1.6"
Cybernator = "^0.6.3"
discord-interactive = "^4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
